+++ test/fixtures/transformation/es6-classes/instance-getter-and-setter/expected.js
@@ -1,13 +1,13 @@
 "use strict";
 
-var _classProps = function (child, staticProps, instanceProps) {
+var _prototypeProperties = function (child, staticProps, instanceProps) {
   if (staticProps) Object.defineProperties(child, staticProps);
   if (instanceProps) Object.defineProperties(child.prototype, instanceProps);
 };
 
 var Test = function Test() {};
 
-_classProps(Test, null, {
+_prototypeProperties(Test, null, {
   test: {
     get: function () {
       return 5 + 5;