+++ test/fixtures/transformation/es6-classes/static/expected.js
@@ -1,6 +1,6 @@
 "use strict";
 
-var _classProps = function (child, staticProps, instanceProps) {
+var _prototypeProperties = function (child, staticProps, instanceProps) {
   if (staticProps) Object.defineProperties(child, staticProps);
   if (instanceProps) Object.defineProperties(child.prototype, instanceProps);
 };
@@ -9,7 +9,7 @@ var A = function A() {};
 
 A.a = function () {};
 
-_classProps(A, {
+_prototypeProperties(A, {
   b: {
     get: function () {},
     set: function (b) {},